requirements.txt
 numpy-1.24.1
 scipy-1.10.0
 math

 
